readline/buffer.go -> buffer.go , puts.go , repaints.go
