+++ readline/buffer.go
@@ -1,49 +1,11 @@
 package readline
 
 import (
-	"fmt"
-	"io"
 	"strings"
 	"unicode"
 
-	"github.com/mattn/go-runewidth"
 	"github.com/mattn/go-tty"
 )
-
-var SurrogatePairOk = false
-
-func (this *Buffer) putRune(ch rune) {
-	if ch < ' ' {
-		this.Out.WriteByte('^')
-		this.Out.WriteByte(byte('A' + (ch - 1)))
-	} else if (ch >= 0x10000 && !SurrogatePairOk) || runewidth.RuneWidth(ch) == 0 {
-		fmt.Fprintf(this.Out, "<%X>", ch)
-	} else {
-		this.Out.WriteRune(ch)
-	}
-}
-
-func (this *Buffer) putRunes(ch rune, n width_t) {
-	if n <= 0 {
-		return
-	}
-	this.putRune(ch)
-	for i := width_t(1); i < n; i++ {
-		this.Out.WriteRune(ch)
-	}
-}
-
-func (this *Buffer) backspace(n width_t) {
-	if n > 1 {
-		fmt.Fprintf(this.Out, "\x1B[%dD", n)
-	} else if n == 1 {
-		this.Out.WriteByte('\b')
-	}
-}
-
-func (this *Buffer) Eraseline() {
-	io.WriteString(this.Out, "\x1B[0K")
-}
 
 const forbiddenWidth width_t = 3 // = lastcolumn(1) and FULLWIDTHCHAR-SIZE(2)
 
@@ -94,10 +56,6 @@ func (this *Buffer) Delete(pos int, n int) width_t {
 	return delw
 }
 
-func (this *Buffer) InsertAndRepaint(str string) {
-	this.ReplaceAndRepaint(this.Cursor, str)
-}
-
 // ResetViewStart set ViewStart the new value which should be.
 // It does not update screen.
 func (this *Buffer) ResetViewStart() {
@@ -116,82 +74,12 @@ func (this *Buffer) ResetViewStart() {
 	}
 }
 
-func (this *Buffer) ReplaceAndRepaint(pos int, str string) {
-	// Cursor rewind
-	this.backspace(this.GetWidthBetween(this.ViewStart, this.Cursor))
-
-	// Replace Buffer
-	this.Delete(pos, this.Cursor-pos)
-
-	// Define ViewStart , Cursor
-	this.Cursor = pos + this.InsertString(pos, str)
-	this.ResetViewStart()
-
-	// Repaint
-	w := width_t(0)
-	for _, ch := range this.Buffer[this.ViewStart:this.Cursor] {
-		this.putRune(ch)
-		w += GetCharWidth(ch)
-	}
-	bs := width_t(0)
-	for _, ch := range this.Buffer[this.Cursor:] {
-		w1 := GetCharWidth(ch)
-		if w+w1 >= this.ViewWidth() {
-			break
-		}
-		this.putRune(ch)
-		w += w1
-		bs += w1
-	}
-	this.Eraseline()
-	if bs > 0 {
-		this.backspace(bs)
-	}
-}
-
 func (this *Buffer) GetWidthBetween(from int, to int) width_t {
 	width := width_t(0)
 	for _, ch := range this.Buffer[from:to] {
 		width += GetCharWidth(ch)
 	}
 	return width
-}
-
-// Repaint buffer[pos:] + " \b"*del but do not rewind cursor position
-func (this *Buffer) Repaint(pos int, del width_t) {
-	bs := width_t(0)
-	vp := this.GetWidthBetween(this.ViewStart, pos)
-
-	for _, ch := range this.Buffer[pos:] {
-		w1 := GetCharWidth(ch)
-		if vp+w1 >= this.ViewWidth() {
-			break
-		}
-		this.putRune(ch)
-		vp += w1
-		bs += w1
-	}
-	this.Eraseline()
-	if del > 0 {
-		this.backspace(bs)
-	} else {
-		// for readline_keyfunc.go: KeyFuncInsertSelf()
-		this.backspace(bs + del)
-	}
-}
-
-func (this *Buffer) RepaintAfterPrompt() {
-	this.ResetViewStart()
-	for _, ch := range this.Buffer[this.ViewStart:this.Cursor] {
-		this.putRune(ch)
-	}
-	this.Repaint(this.Cursor, 0)
-}
-
-func (this *Buffer) RepaintAll() {
-	this.Out.Flush()
-	this.TopColumn, _ = this.Prompt()
-	this.RepaintAfterPrompt()
 }
 
 func (this *Buffer) SubString(start, end int) string {
